refactor: move generation util to util package
